+++ hc/api/models.py
@@ -585,7 +585,14 @@ class Channel(models.Model):
     def discord_webhook_url(self):
         assert self.kind == "discord"
         doc = json.loads(self.value)
-        return doc["webhook"]["url"]
+        url = doc["webhook"]["url"]
+
+        # Discord migrated to discord.com,
+        # and is dropping support for discordapp.com on 7 November 2020
+        if url.startswith("https://discordapp.com/"):
+            url = "https://discord.com/" + url[23:]
+
+        return url
 
     @property
     def discord_webhook_id(self):